requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.3.0
Django==4.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
PyMySQL==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
